backend/requirements.txt
# 深觅 AI Coach - 依赖列表

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据验证
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# AI集成（可选，根据使用的LLM提供商选择安装）
openai>=1.55.3
anthropic==0.18.0
litellm==1.59.8  # 统一LLM路由框架
# lazyllm==0.2.0  # 如果使用LazyLLM框架

# HTTP客户端
httpx==0.26.0
aiohttp==3.9.1

# 数据库
sqlalchemy==2.0.27
aiosqlite==0.19.0

# 认证与安全
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# 工具库
python-dotenv==1.0.0
structlog==24.1.0

# 测试
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # 用于测试

# Mem0 记忆系统
mem0ai>=0.1.21
chromadb>=0.4.18
sentence-transformers>=2.5.0

# 开发工具（可选）
# black==23.12.1
# isort==5.13.2
# flake8==7.0.0
# mypy==1.8.0
